auth/requirements.txt
httpx==1.0.0b0
fastapi==0.110.0
fastapi-limiter==0.1.6
pydantic[email]==2.6.1
pydantic-settings==2.2.1
psycopg2-binary==2.9.9
orjson==3.9.14
python-dotenv==1.0.1
uvicorn==0.27.1
gunicorn==21.2.0

sqlalchemy==2.0.27
alembic==1.13.1
redis==5.0.1
asyncpg==0.29.0

typer==0.9.0
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pyjwt[crypto]==2.8.0
itsdangerous==2.1.2

authlib==1.3.0
requests==2.31.0

grpcio-tools==1.62.0rc1

opentelemetry-api==1.17.0
opentelemetry-sdk==1.17.0
opentelemetry-instrumentation-fastapi==0.38b0
opentelemetry-exporter-jaeger==1.17.0
